refactor: mv slider Input -> SliderPresenter
Rename slider/presenters/Input to slider/presenters/SliderPresenter to
match organization in similar packages.
