+++ packages/slider/src/presenters/SliderPresenter.stylesheet.js
